refactor: Chop template to engine library
