+++ src/games/blackjack.cpp
@@ -1,6 +1,8 @@
 // HEADERS
+#include <iostream>
+
+#include "../engine/engine.h"
 #include "games.h"
-#include "../engine/templates/blackjack.h"
 
 // CONFIG
 int maxHand = 10;
